Move parser interface closer to calling entity
