+++ .github/workflows/run_tests.yml
@@ -19,4 +19,4 @@ jobs:
           go-version: '1.23'
 
       - name: Run tests
-        run: make test
+        run: make test-with-backend
